# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HS FACOP API",
    "description": "HISTORIA CLINICA ORTODONCIA",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/orthodontic-records/{id}": {
      "get": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Obtener una historia",
        "operationId": "getRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historia",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no está a tu alcance",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Guardar la historia completa (con la versión cargada)",
        "operationId": "updateRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Historia guardada, con su nueva versión",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no está a tu alcance",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "La historia cambió desde que se cargó (stale-record) o cambia datos del paciente fijos (patient-locked)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orthodontic-records": {
      "get": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Listar historias (las propias; todas si es ADMIN)",
        "operationId": "listRecords",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Busca en paciente, documento y número (sin distinguir tildes ni mayúsculas)",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "minLength": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Página (base 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Tamaño (1–100; uno mayor se recorta a 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de historias",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRecordSummaryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Página, tamaño o búsqueda inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Crear una historia (borrador) con el siguiente número del autor",
        "operationId": "createRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Historia creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "El tratante llegó a su cupo de historias (record-quota-reached)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/orthodontic-records/{id}/unlock-request": {
      "post": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Solicitar el desbloqueo de los datos del paciente",
        "operationId": "requestPatientUnlock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Solicitud registrada"
          },
          "400": {
            "description": "Motivo inválido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no está a tu alcance",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Datos no fijos (patient-not-locked) o solicitud ya pendiente (unlock-already-requested)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Descartar la solicitud de desbloqueo (solo ADMIN; queda registrado)",
        "operationId": "discardPatientUnlockRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Descartada (o no había)"
          },
          "404": {
            "description": "No existe",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/orthodontic-records/{id}/print": {
      "post": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Registrar una impresión: la primera fija los datos del paciente; devuelve la fecha y el avance",
        "operationId": "printRecord",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Impresión registrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pasos inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no está a tu alcance",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orthodontic-records/quota": {
      "get": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Cupo de historias del usuario autenticado y cuántas creó",
        "operationId": "getRecordQuota",
        "responses": {
          "200": {
            "description": "Cupo y uso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordQuotaResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orthodontic-records/{id}/patient-lock": {
      "delete": {
        "tags": [
          "orthodontic-records"
        ],
        "summary": "Desbloquear los datos del paciente (solo ADMIN; queda registrado)",
        "operationId": "unlockPatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Desbloqueada"
          },
          "404": {
            "description": "No existe",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "409": {
            "description": "Los datos no estaban fijos (patient-not-locked)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso para realizar esta acción",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ValidationProblem": {
        "type": "object",
        "description": "Error de validación (RFC 9457) con detalle por campo.",
        "properties": {
          "type": {
            "type": "string",
            "example": "/errors/validation-error"
          },
          "title": {
            "type": "string",
            "example": "Validation failed"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "Mensaje en español apto para el usuario final."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "instance": {
            "type": "string",
            "example": "/auth/register"
          },
          "traceId": {
            "type": "string",
            "description": "Igual a la cabecera X-Trace-Id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            }
          }
        },
        "required": [
          "detail",
          "status",
          "timestamp",
          "title",
          "type"
        ]
      },
      "ApiProblem": {
        "type": "object",
        "description": "Error de la API (RFC 9457, application/problem+json).",
        "properties": {
          "type": {
            "type": "string",
            "example": "/errors/email-already-exists"
          },
          "title": {
            "type": "string",
            "example": "Conflict"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 409
          },
          "detail": {
            "type": "string",
            "description": "Mensaje en español apto para el usuario final."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "instance": {
            "type": "string",
            "example": "/auth/register"
          },
          "traceId": {
            "type": "string",
            "description": "Igual a la cabecera X-Trace-Id."
          }
        },
        "required": [
          "detail",
          "status",
          "timestamp",
          "title",
          "type"
        ]
      },
...
    }
  }
}